Better test quite naming.
Jenkins show a hierarchical test view by spliting the test name by dots (.)
and in this case the platform and browser version can contain dots that we
do not want to split on. The chosen solution is to xml encode the dot
character in version string and to achieve a nice hierarcy by generating a
test suite naming by joining the test name, platform name and browser name
by dot.